refactor(Formatter): 클래스 이름 변경 + 하드코딩 제거
Before: Util
After: Formatter
