+++ docs/README.md
@@ -16,13 +16,13 @@
 **당첨 번호 생성**
 
 - [x] 6개의 당첨 번호를 입력받는다. - Lotto
-    - [x] 당첨번호 입력값에서 공백을 모두 제거한다. - Util#removeSpace()
+    - [x] 당첨번호 입력값에서 공백을 모두 제거한다. - Formatter#removeSpace()
     - [x] 쉼표를 기준으로 구분한다. - Lotto#separateStringByComma()
     - [x] 구분된 당첨번호의 개수가 6개가 아니면 예외 처리한다. - Lotto#validate()
-    - [x] 구분된 당첨번호가 1~45 사이의 숫자가 아니면 예외 처리한다. - Util#verifyRangeOfLottoNumber()
+    - [x] 구분된 당첨번호가 1~45 사이의 숫자가 아니면 예외 처리한다. - Formatter#verifyRangeOfLottoNumber()
     - [x] 중복되는 숫자가 있으면 예외 처리한다. - Lotto#validate()
 - [x] 보너스 번호 1개를 입력받는다. - Bonus
-    - [x] 보너스 번호가 1~45 사이의 숫자가 아니면 예외 처리한다. - Util#verifyRangeOfLottoNumber()
+    - [x] 보너스 번호가 1~45 사이의 숫자가 아니면 예외 처리한다. - Formatter#verifyRangeOfLottoNumber()
 - [x] 보너스 번호가 당첨번호와 중복되면 예외 처리한다. - LottoController#validateDuplicates()
 
 **로또 결과 계산**
@@ -59,6 +59,7 @@
 
 ## 프로그래밍 요구사항
 
+- [ ] 열 제한은 120자이다. 
 - [ ] 함수(또는 메서드)의 길이가 15라인을 넘어가지 않도록 구현한다.
 - [ ] else 예약어 및 switch/case를 쓰지 않는다.
 - [x] Java Enum을 적용한다.
