+++ src/main/java/lotto/view/InputView.java
@@ -1,19 +1,9 @@
 package lotto.view;
 
-import static lotto.util.Util.removeSpace;
-
 import camp.nextstep.edu.missionutils.Console;
-import java.util.Arrays;
-import java.util.List;
-import java.util.stream.Collectors;
 
 public class InputView {
-
     public static String getInput() {
         return Console.readLine();
     }
-
-
-
-
 }
